+++ R/lavaan_reg.R
@@ -1,6 +1,8 @@
 #' @title Extract relevant regression indices from lavaan model
 #'
-#' @description Extract relevant regression indices from lavaan model.
+#' @description Extract relevant regression indices from lavaan model through
+#'              `lavaan::parameterEstimates` with `standardized = TRUE`. In this
+#'              case, the beta (B) represents the resulting `std.all` column.
 #'
 #' @param fit lavaan fit object to extract fit indices from
 #' @param nice_table Logical, whether to print the table as a `rempsyc::nice_table`