+++ R/save_as_x.R
@@ -2,12 +2,6 @@
 #'
 #' @description These `save_as_x` functions are imported from `flextable`.
 #'
-#' flextable:
-#'       `save_as_docx`,
-#'       `save_as_html`,
-#'       `save_as_image`,
-#'       `save_as_pptx`
-#'
 #' @param ... arguments passed to `save_as_x` functions
 #' @name save_as_x
 NULL